requirements.txt
streamlit==1.35.0
transformers==4.41.0
torch==2.3.0
scikit-learn==1.5.0
shap==0.45.0
fairlearn==0.10.0
aif360==0.6.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
plotly==5.22.0
google-cloud-aiplatform>=1.50.0
google-genai>=1.5.0
openai>=1.30.0
python-dotenv==1.0.1
scipy==1.13.0
accelerate>=0.31.0
datasets==2.19.0
sentence-transformers==2.7.0
fastapi==0.111.0
uvicorn==0.29.0
